reorganize functions with breaking equivalence
